- Fixed table errors - Other minor fixes
