+++ server/src/entities/user/Users.options.ts
@@ -1,15 +1,15 @@
-export const IdOptions: any = {
-    type: 'uuid',
-}
-
 export const UsernameOptions: any = {
     unique: true,
     type: 'varchar',
-    precision: 100,
+    length: 100,
 }
 
 export const EmailOptions: any = {
     unique: true,
     type: 'varchar',
-    precision: 254,
+    length: 240,
+}
+
+export const PasswordOptions: any = {
+    type: 'varchar',
 }
